refactor(message): InputMessage를 message 패키지에 분리
